backend/requirements.txt
pymongo==4.6.1
dnspython==2.4.2
pymongo
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
pyodbc==5.0.1